training/requirements.txt
mlflow==2.13.0
xgboost==2.0.3
scikit-learn==1.4.2
pandas==2.2.0
numpy==1.26.4
psycopg2-binary==2.9.9
redis==5.0.4
pyarrow==15.0.0
SQLAlchemy==2.0.29
imbalanced-learn==0.12.2
fairlearn==0.10.0
aif360==0.6.1
joblib==1.4.0
pyyaml==6.0.1